backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlmodel==0.0.16
sqlalchemy==2.0.35
alembic==1.13.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Encryption
cryptography==42.0.5

# GitHub API
PyGithub==2.1.1

# Templates
jinja2==3.1.4
pyyaml==6.0.1

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.0
python-dotenv==1.0.1
email-validator==2.2.0

# HTTP Client
httpx==0.27.0

# Date/Time
python-dateutil==2.9.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-playwright==0.4.4
playwright==1.48.0
